+++ src/utils/constants.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass, field
 from datetime import datetime
 
 APP_NAME = 'GalePost'
-APP_VERSION = '0.2.13'
+APP_VERSION = '0.2.14'
 APP_ORG = 'GalePost'
 LOG_UPLOAD_ENDPOINT = 'https://galepost.jasmer.tools/logs/upload'